+++ test/lib/Elastica/Test/Query/RescoreTest.php
@@ -200,6 +200,8 @@ class RescoreTest extends BaseTest
 
         $index = $this->_createIndex();
         $index->refresh();
+        $index->optimize();
+
         $results = $index->search($query);
         $response = $results->getResponse();
 
@@ -227,6 +229,8 @@ class RescoreTest extends BaseTest
 
         $index = $this->_createIndex();
         $index->refresh();
+        $index->optimize();
+
         $results = $index->search($query);
         $response = $results->getResponse();